Create a pad for tokens
The idea of the pad is that the grimoire is just a box where tokens can
be added. This does that and allows new tokens to be created and moved
around (although restricting them to the box) and having a click
handler. This will allow a token to be displayed however the storyteller
prefers and they can be clicked on to modify them (change character,
toggle shroud, remove etc.).