Start creating the Select Character functionality
The dialog that allows players to select their characters has been
created. Still to do is making a dialog for the selection and adding it
to the grimoire.

This commit also removes the default styles for the dialog polyfill out
of the main style sheets, only including them if they're needed.