Add a character selection screen
The characters can now be randomly highlighted and added to a new screen
that allows a player (or the story teller on the player's behalf) to
choose a character, which is then added to the grimoire.

This moves the version to 0.3.0-beta - the system should now be usable
in a very basic way.