Save input states and rotate tokens
The input values are now saved so that the full state can be remembered
when the page re-loads. Events are fired so that the game is correctly
set up. This also includes the opened/closed state of the details
sections.

This commit also adds the ability to turn the character tokens
upside-down to indicate that they're the oppose alignment (or show that
a traveller is on the evil team).