Create a new dialog for showing tokens
The new dialog doesn't have any buttons so that a savvy player can't
work out if the token is in play or a demon bluff. This commit also adds
a dialog showing all tokens, allowing a story teller to quickly show a
drunk/poisoned player a not-in-play character, with the player none the
wiser. Crucially, showing a token always takes the same number of taps,
regardless of how it's found, further obscuring the source.

Fixes: #20