Replace the QR code generation library
The old library was larger and slower, and it created the SVG for the QR
code by adding a lot of <path> elements (typically 4,000 - 5,000). This
library is smaller, making the overall file size smaller, and only
generates a single <path> element, which not only speeds up the page,
but makes it much quicker to toggle Travellers and/or Fabled.