Create an "Add traveller" button
The "Add traveller" button allows us to add any traveller to the
grimoire. It only shows travellers that are in the script, but a toggle
switch allows us to see the rest of them. When added to the grimoire,
the traveller's reminder tokens and night order entries are added to the
relevant sections.

This commit also removes the travellers from the other sections, since
they're handled seperately now.

Fixes: #22