Remove the text shadow from the token
There was an annoying bug on the iPhone SE: the text in the token
correctly shrunk down to match the size, but the CSS text-shadow didn't
shrink at the same time. After a bit of trial-and-error, it turns out
that the simplest solution is to simply remove the text shadow. Although
that can make the token text slightly harder to read, it's only
"slightly" and there are other places where the text is more legible.

Fixes: #10