Fix a bug with duplicate tokens
If two tokens were added to the grimoire using the "Add token" or "Add
reminder" lists, only the first one was being stored in the store. This
meant that if either one was moved, only the position of the first one
was remembered. This was a serious hinderence for character like the
Juggler where the same token may need to be added multiple times.

Fixes: #27