Limit the token/info token close chances
The tokens could be hidden by tapping the background, info tokens could
be hidden by tapping anywhere on the screen. This lead to a few akward
story-teller/player interactions where the story teller accidentally
showed the player the entire grimoire (or the website) when they
indented to show information. By changing this so that the closing X has
to be clicked and the background is now inert, this should prevent the
issues from happening again in the future.

Fixes: #28