Add a hash to the data
Because the data always had the same URL, even if it had been updated,
the mobile browser sometimes cached it so clearing the Pocket Grimoire
cache and re-fetching the data didn't give the user the latest data.
This commit adds the hash to the data URLs, but also adds a function
that modified the URLs if necessary. The cache now uses a key, rather
than the URL, so the "characters" data can be stored instead of the
ever-changing URL. This should mean that users always have the latest
data if they clear the cache.