Fix an issue with the Character Sheet QR code
There was a bug where names and character lists were appended to the URL
rather than being set. This caused a weird bug where the name wouldn't
seem to update if the character list was added from the store and then
the edition was changed. Because the name was being appended, it was
sometimes being ignored by the browser. This bug also highlighted that
the name wasn't being removed if a new script was loaded without a name.