Create a homebrew record
The homebrew record is how we're going to allow homebrew characters to
be added by a storyteller and seen by the players. We store the data in
the database and allow players to access it.

This is still a work-in-progress so it hasn't been hooked up or tested
yet, and the class itself has a few notes to remind me what I was
thinking of.

See: #21