Allow custom info tokens to be created
The code that activates the dialogs had to be changed to allow dialogs
to be created on the fly. Generating the default info tokens using the
same technique tests the functionality as well as reducing the markup so
the page loads a little faster.

This is still a work-in-progress (WIP). The custom info tokens still
need to be saved and we need to be able to delete them. I might re-write
the info tokens JavaScript as a class so that the store can pass data a
little easier.

See: #37