Improve custom info tokens
The code for the tokens has been abstracted into a class to make it
easier for other files to add them (such as loading them from the
store). The messages have been translated through Google Translate and
the main info tokens have been replaced with the ones that come in the
physical copy of the grimoire (the colours have been updated to match as
well).

This is still a work-in-progress (WIP) because the custom info tokens
aren't being saved yet.

See: #37