Remove the night order from the character sheet
The official entries don't have the night order, so it makes sense (to
me, at least) that it shouldn't be part of the Pocket Grimoire character
sheet. The code is still in the repo history if we ever want to put it
back.

This also removes the generic icons for the teams because we already
have simple generic icons for them.

See: #48