Tidy the code
The readme file now makes mention of the fact that the Pocket Grimoire
can understand a mixture of homebrew and official characters.

This commit also fixes a bug with the character sheet that prevented the
homebrew game ID being remembered when the language was changed, as well
as replacing the deprecated "addListener" for matchMedia with the
up-to-date addEventListener method.

Fixes: #48
Closes: #52