Prevent multiple tokens being dragged together
The problem was that a second token would be getting dragged while the
first one was still registered, so the event listener wouldn't be
removed and it wouldn't be possible to scroll the grimoire anymore. This
prvents that from happening by cancelling the first token being moved
when the second one is.

Fixes: #66