Add the team breakdown to the character sheet
A new table showing the number of townsfolk, outsiders, minions, and
demons that should be in the game has been added to the character sheet.

This commit also moves the counts into their own YAML file so that the
same data can be displayed in the feed as the table (although I don't
think it will ever change).