Allow for decimals in the night orders
The night orders for homebrew characters can have decimals in their
night orders, which confused the code for adding numbers to the tokens
since it was based on JavaScript arrays. This commit swaps that out for
objects and sorts any keys, allowing decimals of any depth to be
properly handled.

Fixes: #72