Start work on the settable title
The demon bluff groups need to have a settable title to make it easier
for the Story Teller to tell one group apart from another. This is the
same functionality as the token display dialog title, so I'm making a
generic version of that which I can assign to all of them.

This is still a work-in-progress (WIP).

See: #65