Add a note for future work
The demon bluff data in the store will be different from the data that
used to exist. This could cause errors if a user reloads their device
(or re-activates their device) with old data. The note that I've added
will remind me to put in a simple fix for this. This is still a
work-in-progress (WIP).

See: #65