Add the ability to paste a script directly
Because the official script tool now allows the script JSON to be copied
to the clipboard, the Pocket Grimoire should allow people to paste that
directly, rather than having to save it somewhere and then upload it.
This commit does that and updates things like the store to handle the
new textarea field type.

Fixes: #81