Update the night order class
Rather than removing and redrawing all elements in the night order lists
each time a change is made, this update swaps out a comment for the
element (or vice versa) so that the lists can be updated by only
updating the necessary elements.

This commit also sets it so that showing all characters not in play will
make sure that the dead characters are also being shown since this seems
more intuative.

See: #100