Import the new Chinese roles
The roles for "Storm is Coming" (syyl / 山雨欲来) have been imported
and any data issues have been tidied up.

This commit also includes and update to the tokens page that allows the
tokens to clicked to see a larger version of the icon, along with the
ability, id, and edition that the role is from.