Fix a bug with the character ID normalisation
The normalisation hadn't been applied to everything, causing problems
when trying to access the character token. This created defects such as
the inability to select a character from the bag or set it as a demon
bluff. This only applied to characters where the normalised ID was
different from the given ID (i.e. anything with an underscore) so it was
most noticable in homebrew scripts.