Validate the homebrew JSON entries
The homebrew JSON entries are now being validated to makes sure that the
data that's been would successfully create a character. This has mainly
been done so that an error message can be displayed if the data isn't
valid, informing the user before they try to access the character sheet
and get a 500 error.