Highlight a jinx if the correct reminder is used
Some roles should be considered as being added to the grimoire pad if a
reminder is in play, rather than the role itself. Good example of this
are the Drunk and the Marionette. This commit highlights any jinxes that
should occur when this has happened.

This commit also includes information about the homebrew script data,
updating the Readme at the same time.

Fixes: #128