Remove highlight.js
For some reason, highlight.js was causing a problem with the dragging
code - it prevented the user from "letting go" of the token that they
dragged, so they were un able to move any other one. It was a bug that
only affected desktops - mobiles were fine.

Until the reason can be discovered (or resolved) then highlight.js will
remain off the site.

Fixes: #137