Append ellipses to buttons that open dialogue boxes
When using Pocket Grimoire for the first time, I found it wasn’t always
obvious to me which buttons would have some immediate effect related to
the current page, versus which would open a new dialoge box on which I
would perform the action.

If buttons that open dialogue boxes end with the string “…”, I can click
on them without fear that the current state of the grimoire will be
messed up.